Reduces duplicate code in ws-smartmetering response message processors.
