+++ osgp-adapter-ws-smartmetering/src/main/java/com/alliander/osgp/adapter/ws/smartmetering/domain/entities/MeterResponseData.java
@@ -52,16 +52,6 @@ public class MeterResponseData extends AbstractEntity {
 
     }
 
-    // TODO: delete constructor after refactoring complete.
-    public MeterResponseData(final String organisationIdentification, final String messageType,
-            final String deviceIdentification, final String correlationUid, final Serializable messageData) {
-        this.organisationIdentification = organisationIdentification;
-        this.messageType = messageType;
-        this.deviceIdentification = deviceIdentification;
-        this.correlationUid = correlationUid;
-        this.messageData = messageData;
-    }
-
     public MeterResponseData(final String organisationIdentification, final String messageType,
             final String deviceIdentification, final String correlationUid, final ResponseMessageResultType resultType,
             final Serializable messageData) {
