+++ osgp-adapter-ws-smartmetering/src/main/java/com/alliander/osgp/adapter/ws/smartmetering/infra/jms/messageprocessor/ActualMeterReadsResponseMessageProcessor.java
@@ -7,93 +7,15 @@
  */
 package com.alliander.osgp.adapter.ws.smartmetering.infra.jms.messageprocessor;
 
-import javax.jms.JMSException;
-import javax.jms.ObjectMessage;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.alliander.osgp.adapter.ws.schema.smartmetering.notification.NotificationType;
-import com.alliander.osgp.adapter.ws.smartmetering.application.services.MeterResponseDataService;
-import com.alliander.osgp.adapter.ws.smartmetering.application.services.NotificationService;
-import com.alliander.osgp.adapter.ws.smartmetering.domain.entities.MeterResponseData;
 import com.alliander.osgp.domain.core.valueobjects.DeviceFunction;
-import com.alliander.osgp.domain.core.valueobjects.smartmetering.MeterReads;
-import com.alliander.osgp.domain.core.valueobjects.smartmetering.MeterReadsGas;
-import com.alliander.osgp.shared.infra.jms.Constants;
-import com.alliander.osgp.shared.infra.jms.ResponseMessageResultType;
 
 @Component("domainSmartMeteringActualMeterReadslResponseMessageProcessor")
 public class ActualMeterReadsResponseMessageProcessor extends DomainResponseMessageProcessor {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(ActualMeterReadsResponseMessageProcessor.class);
-
-    @Autowired
-    private NotificationService notificationService;
-
-    @Autowired
-    private MeterResponseDataService meterResponseDataService;
 
     protected ActualMeterReadsResponseMessageProcessor() {
         super(DeviceFunction.REQUEST_ACTUAL_METER_DATA);
     }
 
-    @Override
-    public void processMessage(final ObjectMessage message) throws JMSException {
-        LOGGER.debug("Processing smart metering actual meter reads response message");
-
-        String correlationUid = null;
-        String messageType = null;
-        String organisationIdentification = null;
-        String deviceIdentification = null;
-
-        String notificationMessage = null;
-        NotificationType notificationType = null;
-        ResponseMessageResultType resultType = null;
-
-        try {
-            correlationUid = message.getJMSCorrelationID();
-            messageType = message.getJMSType();
-            organisationIdentification = message.getStringProperty(Constants.ORGANISATION_IDENTIFICATION);
-            deviceIdentification = message.getStringProperty(Constants.DEVICE_IDENTIFICATION);
-            resultType = ResponseMessageResultType.valueOf(message.getStringProperty(Constants.RESULT));
-
-            notificationMessage = message.getStringProperty(Constants.DESCRIPTION);
-            notificationType = NotificationType.valueOf(messageType);
-
-        } catch (final JMSException e) {
-            LOGGER.error("UNRECOVERABLE ERROR, unable to read ObjectMessage instance, giving up.", e);
-            LOGGER.debug("correlationUid: {}", correlationUid);
-            LOGGER.debug("messageType: {}", messageType);
-            LOGGER.debug("organisationIdentification: {}", organisationIdentification);
-            LOGGER.debug("deviceIdentification: {}", deviceIdentification);
-            return;
-        }
-
-        try {
-            LOGGER.info("Calling application service function to handle response: {}", messageType);
-
-            if (message.getObject() instanceof MeterReads) {
-                // Convert and Persist data
-                final MeterReads data = (MeterReads) message.getObject();
-                final MeterResponseData meterResponseData = new MeterResponseData(organisationIdentification,
-                        messageType, deviceIdentification, correlationUid, resultType, data);
-                this.meterResponseDataService.enqueue(meterResponseData);
-            } else {
-                final MeterReadsGas meterReadsGas = (MeterReadsGas) message.getObject();
-                final MeterResponseData meterResponseData = new MeterResponseData(organisationIdentification,
-                        messageType, deviceIdentification, correlationUid, meterReadsGas);
-                this.meterResponseDataService.enqueue(meterResponseData);
-            }
-
-            // Send notification indicating data is available.
-            this.notificationService.sendNotification(organisationIdentification, deviceIdentification,
-                    resultType.name(), correlationUid, notificationMessage, notificationType);
-
-        } catch (final Exception e) {
-            this.handleError(e, correlationUid, organisationIdentification, deviceIdentification, notificationType);
-        }
-    }
 }
